CCD-93: Added starting/leaving dates and availability to the new temporary endpoint team2 which is meant to replace the endpoint team.
The JIRA fields dateTo, dateFrom and availability are now stored in an additional team_member table. The team table has a members column with a map containing usernames and starting dates. Unfortunately, a map with a single date value is no longer sufficient. Currently, the team_member table and the members column of the team table redundantly store the dateFrom value, however this hasn't been changed deliberately to preserve downwards compatibility as long as the team endpoint has not been fully replaced by the (current) team2 endpoint.